Added Symbol parameter for hooks
Symbols use the "symb" parameter to read the offset of a function in the .sym file.
adding "addr" will offset the symbol so it's "symbol + offset".
This is useful for games like Fates that ship with full symbols so hooks are harded coded to an offset and allow dynamic hooking into code.